fixed githubIds type in db

+++ database/migrations/2026_07_28_203155_create_github_repositories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('github_repositories', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('github_id')->unique();
+            $table->bigInteger('github_id')->unique();
             $table->string('name');
             $table->string('full_name');
             $table->string('url');
